Разбивает на модули
