+++ webpack.config.js
@@ -2,12 +2,12 @@ const path = require(`path`);
 
 module.exports = {
   mode: `development`,
-  entry: `./src/main/js`,
+  entry: `./src/main.js`,
   output: {
     filename: `bundle.js`,
     path: path.join(__dirname, `public`)
   },
-  devtool: `sourse-map`,
+  devtool: `source-map`,
   devServer: {
     contentBase: path.join(__dirname, `public`),
     watchContentBase: true
